+++ workspaces/scorecard/plugins/scorecard-common/src/index.ts
@@ -13,25 +13,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/**
- * Common functionalities for the scorecard plugin.
- *
- * @packageDocumentation
- */
 
-/**
- * In this package you might for example declare types that are common
- * between the frontend and backend plugin packages.
- *
- * @public
- */
-export type CommonType = {
-  field: string;
-};
-
-/**
- * Or you might declare some common constants.
- *
- * @public
- */
-export const COMMON_CONSTANT = 1;
+export * from './types';